Installing airspeed based on OS.

+++ init.sh
@@ -24,8 +24,12 @@ if [ $? ]; then
 fi
 cd ..
 echo "installing airspeed dependency"
-git clone https://github.com/purcell/airspeed.git
-cd airspeed
-sudo python3 setup.py install
-cd ..
-sudo rm -Rf airspeed
+if [[ "$OSTYPE" == "darwin"* ]]; then
+	pip3 install airspeed
+else
+	git clone https://github.com/purcell/airspeed.git
+	cd airspeed
+	sudo python3 setup.py install
+	cd ..
+	sudo rm -Rf airspeed
+fi
